requirements.txt
brevitas==0.10.2
certifi==2025.7.14
charset-normalizer==3.4.2
coloredlogs==15.0.1
concrete-ml==1.9.0
concrete-ml-extensions==0.1.9
concrete-python==2.10.0
dependencies==2.0.1
dill==0.4.0
filelock==3.18.0
flatbuffers==25.2.10
fsspec==2025.7.0
hf-xet==1.1.5
huggingface-hub==0.33.4
humanfriendly==10.0
hummingbird-ml==0.4.11
idna==3.10
importlib_resources==6.5.2
Jinja2==3.1.6
joblib==1.5.1
jsonpickle==4.1.1
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.2.1
numpy==1.26.4
onnx==1.17.0
onnxconverter-common==1.15.0
onnxmltools==1.14.0
onnxoptimizer==0.3.13
onnxruntime==1.18.0
packaging==25.0
pandas==2.0.3
protobuf==3.20.3
psutil==7.0.0
python-dateutil==2.9.0.post0
pytz==2025.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.5.0
scipy==1.13.1
six==1.17.0
skl2onnx==1.19.1
skops==0.5.0
skorch==0.11.0
sympy==1.14.0
tabulate==0.9.0
threadpoolctl==3.6.0
tokenizers==0.21.2
torch==2.3.1
tqdm==4.67.1
transformers==4.53.2
typing_extensions==4.14.1
tzdata==2025.2
unfoldNd==0.2.3
urllib3==2.5.0
xgboost==1.6.2
z3-solver==4.13.0.0
zipp==3.23.0
